Refactored overwatch and reboot functions out of MatrixFlow
