+++ src/module/flows/MatrixFlow.ts
@@ -4,7 +4,6 @@ import { Helpers } from '../helpers';
 import { MatrixNetworkFlow } from '../item/flows/MatrixNetworkFlow';
 import { SR5Item } from '../item/SR5Item';
 import { MatrixRules } from '../rules/MatrixRules';
-import { MarksStorage } from '../storage/MarksStorage';
 import { OpposedTest } from '../tests/OpposedTest';
 import { SuccessTest } from '../tests/SuccessTest';
 import { DamageType } from '../types/item/Action';
@@ -58,74 +57,6 @@ export const MatrixFlow = {
         };
 
         return updateData;
-    },
-
-    /**
-     * Determine if this test should add overwatch score. SR5#231-232 'Overwatch Score and Convergence'
-     *
-     * @param test Any test, though only tests opposing matrix tests will be considered.
-     */
-    async addOverwatchScoreFromIllegalMatrixAction(test: SuccessTest | OpposedTest) {
-        if (!test.opposing) return;
-        // @ts-expect-error - Only OpposedTest has this property
-        const against = test.against as SuccessTest;
-        if (!against) return;
-        if (!against.hasTestCategory('matrix')) return;
-        if (!MatrixRules.isIllegalAction(
-            against.data.action.attribute as any,
-            against.data.action.attribute2 as any,
-            against.data.action.limit.attribute as any)) {
-            return;
-        }
-
-        const actor = against.actor;
-        if (!actor) return;
-
-        // Raise overwatch score.
-        let overwatchScore = actor.getOverwatchScore();
-        await actor.setOverwatchScore(overwatchScore + test.hits.value);
-
-        // Inform GM about convergenace.
-        overwatchScore = actor.getOverwatchScore();
-        if (MatrixRules.isOverwatchScoreConvergence(overwatchScore)) {
-            await MatrixFlow.executeOverwatchConvergence(actor);
-        }
-    },
-
-    /**
-     * Execute a convergence with the given actor based on SR5#231-232 'Overwatch Score and Convergence'.
-     *
-     * @param actor The actor targeted by Convergence.
-     */
-    async executeOverwatchConvergence(actor: SR5Actor) {
-        // Prepare speaker data.
-        const alias = game.user?.name;
-        const linkedTokens = actor.getActiveTokens(true) || [];
-        const token = linkedTokens.length === 1 ? linkedTokens[0].id : undefined;
-
-        // Prepare convergence damage
-        const damage = MatrixRules.convergenceDamage();
-
-        const resultActions = [
-            {
-                label: 'SR5.Labels.Action.ForceReboot',
-                action: 'forceReboot',
-                value: 'Test',
-            },
-        ];
-
-        const templateData = {
-            overwatchThreshold: MatrixRules.overwatchConvergenceScore(),
-            damage,
-            resultActions,
-            speaker: {
-                actor,
-                alias,
-                token,
-            },
-        };
-
-        await MatrixFlow.sendOverwatchConvergenceMessage(templateData);
     },
 
     async determineMatrixFailedAttack(test: SuccessTest | OpposedTest) {
@@ -202,67 +133,7 @@ export const MatrixFlow = {
         }
     },
 
-    /**
-     * Send out a chat message to inform the GM about convergenace and provide actions for it.
-     */
-    async sendOverwatchConvergenceMessage(templateData) {
-        const content = await renderTemplate(
-            'systems/shadowrun5e/dist/templates/chat/overwatch-convergence-message.hbs',
-            templateData,
-        );
-        const messageData = { content, speaker: templateData.speaker };
-        await ChatMessage.create(messageData);
 
-    },
-
-    /**
-     * Reboot this device and trigger all resulting effects of it based on
-     * the reboot action.
-     * See SR5#242 'Reboot Device'.
-     *
-     * @param device The device to reboot.
-     * @param delay The delay duration of combat turns after which the devices has rebooted.
-     */
-    async rebootPersona(actor: SR5Actor, delay: number = 1) {
-        console.debug('Shadowrun 5e | Rebooting Persona Device', actor);
-        // TODO: Mark devices as rebooting to prohibit usage until end of next combat turn (effect with duration).
-        // TODO: Allow for user input delay when rebooting device
-
-        // Link Locked actors can´t be rebooted. (SR5#229 ´Dumpshock & Link-Locking´)
-        if (actor.isLinkLocked) {
-            ui.notifications?.error('SR5.Matrix.Error.LinkLockedReboot');
-            return;
-        }
-
-        // Only reset overwatch score if the device rebooted is the persona device.
-        const device = actor.getMatrixDevice();
-
-        await actor.setOverwatchScore(0);
-        await actor.clearMarks();
-        await MarksStorage.clearRelations(actor.uuid);
-        const damage = MatrixFlow.getDumpshockDamage(actor);
-
-        await MatrixFlow.sendRebootDeviceMessage(actor, device, delay, damage);
-    },
-
-    /**
-     * Inform GM and users about the result of a reboot action.
-     */
-    async sendRebootDeviceMessage(actor: SR5Actor, device: SR5Item | undefined, delay: number, damage: DamageType) {
-        const speaker = {
-            actor,
-            alias: game.user?.name,
-            token: actor.getActiveTokens(true)[0]?.id,
-        };
-        const content = await renderTemplate('systems/shadowrun5e/dist/templates/chat/reboot-device-message.hbs', {
-            speaker,
-            delay,
-            device,
-            damage
-        });
-        const messageData = { content };
-        await ChatMessage.create(messageData);
-    },
 
     /**
      * Get the amount of damage an actor needs to resist as Biofeedback Damage
